RR - [MOD]: Modifications related to Approval request delete.

+++ setu_pharma_basic/models/tour_plan.py
@@ -48,7 +48,8 @@ class TourPlan(models.Model):
     approval_user_status = fields.Selection(related='approval_request_id.user_status')
     requester_id = fields.Many2one("res.users", string="Requester",
                                    default=lambda self: self.env.user)
-    approval_request_id = fields.Many2one('approval.request', "Approval Request")
+    approval_request_id = fields.Many2one('approval.request', "Approval Request",
+                                          ondelete='cascade')
     division_id = fields.Many2one("setu.pharma.division", string="Division", copy=False,
                                   default=lambda self: self.env.user.employee_id.division_id)
     period_id = fields.Many2one("setu.pharma.fiscalperiod", string="Fiscal period", domain=current_next_fiscal_period,
@@ -293,10 +294,13 @@ class TourPlan(models.Model):
             if tp.search([('period_id', '=', tp.period_id.id), ('id', '!=', tp.id),
                           ('state', '!=', 'cancel')]):
                 raise ValidationError(
-                    _("You can't create multiple tourplan for same fiscal period"))
+                    _("You can't create multiple tour plan for same fiscal period"))
 
     def unlink(self):
         """ Unlink Inherited. """
         if self.state == 'approved':
             raise ValidationError(_("You can't delete 'Approved' Tour Plan"))
+        if self.approval_request_id:
+            raise ValidationError(_("You can't delete Tour Plan having Approval Request. You can "
+                                    "cancel it."))
         return super(TourPlan, self).unlink()
